Turn the acquisition process into a loop and unify end points.
The acquisition process handles endpoint conditions inconsistently,
making the code harder to follow and maintain. Refactor the acquisition
process into a loop with a unified endpoint check to simplify control
flow and make the end of acquisition easier to handle.

Due to an early return, detector state goes to idle rather than aborted
when the ADImageSingle is in use. To keep this behavior without this
early return, check if the imageMode is ADImageMultiple (that is, not
ADImageSingle nor ADImageContinuous) to change to aborted; otherwise, go
to idle.