Turn the non acquisition process into a loop.
Refactor the process into a loop to improve clarity and simplify
handling the start of acquisition.

Note that now it doesn't check for !_exiting anymore, what shouldn't be
a problem since Shutdown() won't wait for any threads at all, and the
thread itself doesn't seem to require terminating in any specific state.